i18n: add canary checking to pot file extraction
In LP#1758684 we got a report that the pot file extraction is
broken. This is not (yet) locally reproducible. Until then this
PR adds canary strings that must be available in the generated
pot.

The example pot files that were broken on LP were very small and
did not had these strings.