tests: do not mask errors in interfaces-timezone-control
We currently mask errors from `timedatectl set-timezone`. This
is bad because:
a) The bug is supposed to be fixed (LP: 1650688)
b) We run `timedatectl set-timezone` in restore and if it fails
   there the entire suite aborts

This PR changes this so that we don't mask the error in the
execution but do not fail in restore anymore (to not block the
entire suite).