i18n: use xgettext-go --files-from to avoid running into cmdline size limits
When xargs gets a commandline that is too long it will spawn the
given tool multiple times. This is not desirable in the xgettext
case as each run overrides the output of the previous run.

So first use "find" to find all the go files we need to look for
translations and then use the new "xgettext-go --from-file" feature
to read the files from. This avoids that we run into limits of
argv again.