cmd/snap-confine: fix nvidia support under lxd
From mount(2), we can learn that MS_REMOUNT has a special meaning when
coupled with MS_BIND:

   Since Linux 2.6.26, this flag can be used with MS_BIND to modify only the
   per-mount-point flags.  This is particularly useful for setting or clearing the
   "read-only"  flag  on  a  mount  point  without changing the underlying
   filesystem.  Specifying mountflags as:

       MS_REMOUNT | MS_BIND | MS_RDONLY

   will make access through this mountpoint read-only, without affecting other
   mount points.

Under LXD snap-confine runs under stacked apparmor profile where the effective
profile is an intersection of the LXD profile and the snap-confine profile. LXD
has a very carefully crafted profile that limits unsafe operations. One of
those is a filesystem remount operation, which could potentially affect the
host. Since this is disallowed by LXD, snap-confine cannot perform the remount
and application startup fails with the following denial.

    [61955.378584] audit: type=1400 audit(1530126985.909:482):
    apparmor="DENIED" operation="mount" info="failed flags match" error=-13
    profile="lxd-brave-drum_</var/snap/lxd/common/lxd>"
    name="/tmp/snap.rootfs_o6dcBz/var/lib/snapd/lib/vulkan/" pid=12945
    comm="snap-confine" flags="ro, remount"

Since this issue is only evident when using nvidia hardware it went under the
radar.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>