tests/lib: sync the file before checking its contents.
This should not work. POSIX demands that it not work. There is a basic
"if you write something to the file, and then read it, you should read
what you just wrote" expectation that would be being broken, so if
this worked a _lot_ of things out there would come into weird issues
and hard-to-debug crazy stuff would be going on all the time.

And yet...