Check refresh before and after restart snapd
As during the prepare we are starting snapd service just before set the
cursor to prevent the log of snaps of previous test is included in the
current journal, it is possible that the "Catalog refresh" is logged
before the cursor is set.

It is easy to reproduce on boards such as pi2 or pi3.

The change is simple an checks the catalog is refreshed after the
restart which is done as part of the test.