tests: "snap connect" is idempotent so just connect
This patch changes the interfaces-many test so that it avoids a check
if a specific interface is connected. First of all using MATCH there
just adds noisy output that looks like a problem while not being one.
Second of all connect is idempotent so this is not something we need
to check for.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>