overlord,daemon: remap connections in {get,set}Conns
This patch moves the logic of re-mapping connections to the edge of
state interaction and daemon API interaction. That is, incoming API
requests and state reads are re-mapped (via incoming mapper) and the
outgoing API responses and sate writes are re-mapped (via outgoing
helper) but there are no other re-mappings in any particular place.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>