overlord/ifacestate: add re-mapping function for plugs and slots
This patch adds re-mapping functions for plugs and slots. The idea with
re-mapping is that snapd can provide the appearance of a particular plug
and slot placement as observed in the on-disk state and over-the-wire
communication while having a different actual placement in memory. In
practice this is designed to allow us to place implicit slots on the
snapd snap.

Re-mapping is implemented by a golang interface. Currently only a no-op
mapper is used. The mapper is global but could be moved to the interface
manager after changing a large number of functions to become methods on
the interface manager. This can be done in a follow-up, if desired.

Apart from the golang interface, which provides the low-level
operations, a number of helpers for other data types are provided. This
includes connection references and plug and slot information objects.
Those helpers are used by the daemon API layer in one of the subsequent
patches.

Two real mappers are implemented: a no-op mapper and a core<=>snapd
mapper. For testing a upper<=>lower case mapper is used, for extra
clarity. Everything is accompanied by unit tests.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>