daemon: re-map interface state in API requests/responses
This patch changes interface connection, disconnection and query APIs to
apply re-mapping on incoming requests and outgoing responses. This
allows code to issue explicit connect and disconnect API requests for
the "core" snap even though implicit slots are on the "snapd" snap.
Similarly the API for querying interface connections (both the legacy,
as used by "snap interfaces" and the per-interface as used by "snap
interface") are updated to take advantage of this feature.

Note that at this stage the code is still dormant since re-mapping is a
no-op until the "snapd" snap becomes the host of the implicit slots.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>