tests: don't rely on sleep in service-stop-mode tests
The tests were flaky and required a "sleep 1" workaround because the
service was considered ready before the running service process was
able to register signal handlers. This can be properly fixed with
"daemon: notify" and sd_notify but given the fact that the
"daemon-notify" interface doesn't auto-connect it cannot be used for
this purpose at present.

As a workaround the service writes a "ready" file after establishing the
signal handlers and the test waits for the file to show up before
proceeding.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>