snap,snap-exec: support command-chain for app
When a snap is created by the snapcraft CLI, each app has an associated
chunk of environment variables defined as well as scripts that may be
run before the app's command. Currently, this is done by generating a
shell script wrapper for each app, but this is opaque and confusing, and
makes `snap run --shell` not actually provide a representative
environment for the app.

Snapd supports an `environment` keyword for apps and hooks, but the last
item needed before the shell wrappers can go away is the ability to run
scripts before the app command. Add support for this to apps. Hook
support will be a follow-up.

See forum topic https://forum.snapcraft.io/t/6112 for more details.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>