daemon,api: separate mappings for state and API
This patch changes a number of things, the main concept is that
the mapping system can now offer separate replacements for state
handling, where we want to stick to the word "core" for compatibility,
and the API layer where we want to start using the "system" word
as a nickname to whatever is providing implicit slots.

As such the mapping functions are duplicated, once for state and once
for request/response. The intermediate helpers for working with
connection references are dropped because this just multiplies trivial
functions that don't need to be used.

The existing NilMapper is expanded to cover the new API. The existing
CoreSnapdMapper is split into two new mappers: CoreSnapdSystemMapper
which uses core in the state, snapd in memory and system in the API layer
and a new CoreCoreSystemMapper which uses core in the state and in memory
but system in the API layer.

That last new mapper is a candidate for transition of spread tests
that could start to talk about "system" natively (without the hacks in
the client) and would then work without change once the core-snapd-system
mapper can be used, which will happen once snapd starts to host implicit
slots.

The daemon and state are obviously updated to switch to the appopriate
mapping helper.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>