cmd,overlord,daemon: merge snap nickname with iface mapper
While testing the interface mapper I realized the nickname system is
still in use and actually clashes with the interface mapper. Looking at
it close it is obvious that the nickname system must be folded into the
interface mapper for consistency and that the interface mapper must be
able to map snap names more than than plug or slot references.

Without this change we can never have implicit slots on the "snapd" snap
because implicit gadget connections would resolve to "core" (which can
be absent) and because the nickname handling is would clash with the
idea of using "system" in the API layer but "snapd" in the memory
(because nicknames would effectively change "system" to "core".

To fix this I extended the interface mapper to also be a snap name
mapper and removed the nickname system entirely. This also allows us to
drop the ever-growing list of special cases in the client (for interface
abbreviation logic).

More precisely this patch:
 - drops the {Use,Drop}Nick functions from the snap package
 - changes the default mapper to CoreCoreSystem mapper
 - adds a SnapMapper that has request/response state save/load functions
 - makes InterfaceMapper a superset of SnapMapper
 - refactors CoreCoreSystemMapper and CoreSnapdSystemMapper
 - makes the daemon use RemapSnapFromRequest instead of the nick
 - makes gadget resolver use RemapSnapFromRequest to resolve "system"
 - makes the client abbreviate just "system" slots

This comes without a full array of tests but without breaking any
existing test (woot). It also passes spread which is promising.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>