daemon: move to use a constructor for Meta
Before this change, Meta was being built 'by hand' every time it was
needed.  In warnings there's a bit of work to be done every time, so
having a constructor would make things easier. However doing the
change to have a constructor makes that diff unnecessarily large.

So this change creates a newMeta constructor, that you should call
with a locked state. I've changed all uses to do so, but nothing
checks that you do this yet, as without warnings it'd be
pointless. But I'm reworking warnings to use this and that should be
coming up right after.