logger: try to not have double dates
Before this change, logger was inconditionally adding a timestamp to
our logs. Then systemd picks them up and adds a timestamp. This
results in silly logs like

    Jul 19 16:40:30 fleet snapd[8498]: 2018/07/19 16:40:30.298179 retry.go:52: DEBUG: <blah blah blah>

With this change we take advantage of the fact that systemd services
run with a very limited environment, and check for TERM being set: if
not set, we disable logging the timestamp ourselves. That way you can
still run snapd by hand to debug things and get timestamps, but let
systemd do it itself otherwise.