tests: remove `-o pipefail` again
This will fix the spread MATCH breaking we see in out tests. It
also seems like the side-effects from `-o pipefail` are subtle
and confusing. Like `echo "foo\nfoo\nfoo"|grep foo` can break
echo with SIGPIPE and abort the script. So we trade one set of
bugs (early bits in the pipeline fail and we don't notice)
for another set of bugs that is arguable more subtle.