tests/lib/snaps: avoid using relative command paths that go up in the directory tree
Use of relative command paths that go outside of $SNAP tree is problematic and
may produce unexpected absolute paths.

In an example case of AMZN2, the distro ships /etc/os-release as a file (not a
symlink to /usr/lib/os-release like other distros do). Because of this, when the
mount namespace is set up, the /etc/os-release files appears directly in the
namespace, but contains host specific information, i.e ID=amzn,
ID_LIKE="..rhel..". This confuses snap-exec as to the snap mount point location,
which is now /var/lib/snapd/snap instead of /snap (as would be in case when
distro is Ubuntu Core in the 'core' snap).

In case of the test snaps affecte by this patch, the paths ../../../bin/sh ended
up being /var/lib/snapd/bin/sh instead of /bin/sh, thus preventing the snaps
from working correctly.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>