seccomp: conditionally add socketcall() based on system
The socketcall() syscall is an older multiplexing syscall that is historically
used on some architectures (eg, i386) for various socket related operations
like socket(), bind(), listen(), etc. socketcall()'s 'args' cannot be mediated
with seccomp, so not allowing the syscall where possible is desirable. The
current implementation is fine on architectures that don't support socketcall()
at all (eg, amd64, ARM), but socketcall() should only be add conditionally on
the others since it can be used instead of the mediated syscalls. LP: #1446748.

For i386, the 4.3 kernel added individual syscalls for all the multiplexed
socketcall() operations and glibc was modified to prefer the individual
syscalls over socketcall() (statically compiled programs need to have been
built with the newer glibc). With a new enough kernel and glibc, on i386 we
can remove socketcall() since glibc will try the regular syscalls first, then
fallback to socketcall(), thus allowing proper mediation for this family of
syscalls.

As such, snapd will determine if the system requires the use of socketcall().
Factors:
- if the kernel architecture is amd64, armhf or arm64, do not require
  socketcall (unused on these architectures)
- if the kernel architecture is i386:
  - if the kernel is < 4.3, force the use of socketcall()
  - for backwards compatibility, if the system is Ubuntu 16.04 or lower,
    force use of socketcall()
- if the kernel architecture is not any of the above, force the use of
  socketcall() (TODO: verify other architectures)

TODO: in addition to the above, a future commit will consider the specified
base snap (since it might use an older glibc).