devicestate: only run device-hook when fully seeded
We currently run the "seeding" and "become-operational" changes in
parallel. This is fine most of the time. However if the prepare-device
hook is used it may run before the system is fully seeded. This can
be problematic in the general case. It is especially problematic on
core18 devices that will use snap-confine to setup the environment
based on the installed core18 snap. If that snap is not available
yet things will fail.