image: do simple seed.yaml snap sorting
We ran into some issues with projects that had the order in the
seed.yaml wrong. I.e. installing snaps that require content
providers before the content provider. This PR adds a very simple
sorting for seed.yaml:

1. base/core,kernel,gadget,"snapd" are never reordered
2. other bases
3. snaps that provide any content interface slot
4. other snaps

It will also warn if any of the content provider themself use a
content plug that this nees manual resolving.

The above sorting should be fine because:
- we order the boot base/core/gadget/kernel manually already
  so (1) should be fine
- bases can not depend on bases so (2) should be fine
- snaps that provide content must go before snaps that use content,
  this is why we have (3)
- then remaining snaps have any needed bases or content providers
  installed at this point

One caveat for (3) is of course any snaps that both provide and
require content. This case is not handled just warned about.