overlord/devicestate: have the serial request talk to the proxy if set
Without this change, the serial request ignored proxy.store set on
core, meaning it tried to connect directly instead of going through
the proxy. [lp:1781443] is an instance of the problems caused by this.

This change fixes that, and also prepares things so that we can
support talking to the proxy when a gadget sets device-service.url (a
device with device-service.url set on the gadget *and* proxy.store set
on core will today attempt talking directly to device-service.url).

[lp:1781443]: https://bugs.launchpad.net/snapd/+bug/1781443