overlord/devicestate: DTRT w/a snap proxy to reach a serial vault
Without this change, if `proxy.store` was set on core, and
`device-service.url` was set on the gadget, snapd would attempt to
talk directly to the device service.

This change adds the logic so that, in this case, we talk to the
appropriate endpoint of the proxy, specifying the serial vault we want
in a header.