overlod/ifacestate: always use a new SnapState when fetching the snap state
The state is serialized to JSON. When unmarshalling to a struct, the fields not
appearing in the original JSON input will not be assigned to. In this case, the
SnapState was filled with some data from the first .Get() call. When JSON input
was unmarshalled in the second .Get() call, the input data did not have
instance-key set and the resulting snapstate data retained the instance key from
the first call.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>