overlord/devicestate: only do the proxy+serial dance if the proxy is new enough
Without this we'd always ask the proxy for the serial if it and the
vault were set, and older proxies won't understand the request and
hand out a generic serial instead. With this change, if the prox and
the serial vault are both set, we check the proxy store's version and
only do the full proxy+serial dance if it's new enough.