release: detect when apparmor is available but not usable
Under some configuration apparmor may look like its available but
we get permission denied errors when trying to use it. This can
happen on e.g. an lxd container that runs with:

    raw.lxc:
      lxc.apparmor.profile=unconfined

In this case lxd will not setup apparmor stacking so the container
looks unconfined however lxd will not grant CAP_MAC_ADMIN to the
container (which is quite sensible). But it means that snapd will
not be able to setup any apparmor profiles inside such containers.

When this is detected snapd will run just like on any other system
that does not have apparmor confinement (which is probably fine given
that the admin asked the container to run without confinement).

The approach that uses apparmor_parser inside snapd is noisy
right now, it will create an audit message in the journal on
each snapd startup.