advise-snap: add --dump-db which dumps the command database
This is a hook needed for my new command-not-found implementation[1], which
was written to be fast, and remove the dependency on Python.

Currently Ubuntu is running two programs (one which requires an
interpreter) and opening two huge databases
for every command_not_found_handler invocation.

I also will need a hook somewhere so snap can call "update-command-not-found"
whenever it updates its command database.

[1] https://github.com/shawnl/command-not-found