interfaces/builtin: addtl network-manager resolved DBus fix
The previous commit was based on resolvd changes made in
commit 4b73774a which added a stanza for the resolved method's
Resolve{Address,Hostname,Record,Service}, however the peer as
specified (name="org.freedesktop.resolve1") doesn't work for
the case where NetworkManager called SetLinkDNS or SetLinkDomains.
In this case, the peer needs to be specified as (label=unconfined)
for the methods to work properly.