interfaces: add audio-playback and audio-record
The audio-playback and audio-record interfaces are companion interfaces
such that the design of them is based on the idea that the slot
implementation (eg pulseaudio) is expected to query snapd on if the
audio-record slot is connected or not and the audio service will mediate
recording. Since pulseudio is accessed through the same mechanism for
playback or record, the security policy for record doesn't include
anything more (audio recording without playback is intentionally
unsupported). If other audio recording servers require different
security policy for record and playback (eg, a different socket path),
then those access will be added to this interface.

Since playback is deemed safe but recording is not, audio-playback is
auto-connected where audio-record is not.