packaging/opensuse: fix static build of snap-update-ns and snap-exec
snap-update-ns and snap-exec must be built statically. Turns out those were not
build as such on openSUSE. This has worked by accident until now. Once we
started using C library calls such as islower(), isdigit() things started to
blow up and we ended up getting a segfault in the linker.

The patch enables snap-update-ns and snap-exec to be built statically.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>