cmd/libsnap: unify detection of core/classic with go
This patch changes sc_classify_distro to use the same trick that
dirs/dirs.go is using, namely that if the root filesystem contains
/meta/snap.yaml then we are inside a base (nee os) snap and then the
os-release file plays secondary role (it is only used to determine
core16 vs any other base snap).

The same patch fixes unit tests to actually test the variant
os_release_classic_with_long_line (making variables static helps), adds
mocking helpers and switches the tests to use them.

This makes actual tests shorter and easier to follow, we also correctly
both unlink temporary files _and_ restore the value of internal
variables that are mocked for testing.

Lastly the patch adds a new test that explicitly shows that a "custom"
distribution is classified as classic when /meta/snap.yaml is absent but
as core when that file is present.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>