store: backward compatible instance-key handling for non-instance snaps (#5667)
* store: backward compatible instance-key handling for non-instance snaps

With parallel installation of snaps, the snap instance key, as used for cross
referencing the current snaps and actions in store responses is derived at
runtime. This has store side implications and may negatively affect the store.
Revert to the old behavior for snaps that do not have instance key set (i.e. are
named <snap> instead of <snap>_<instance>).

Also, we want the instance-key as sent to store to be stable accross subsequent
requests. Reverting to the old behaviour for plain snaps ensures this property.
Leave a TODO with a note for parallel installed snaps.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>