overlord: handle sigterm during shutdown better
The current snapd reboot hangs in the following scenario:

0. cmd/snapd/main.go runs and sets up a SIGTERM handler
1. snapd refreshes core
2. this triggers daemon:reboot() which will initiate a system shutdown
3. daemon.Start() exits
4. daemon.Stop() runs and waits for a reboot
5. systemd sends a SIGTERM to snapd as part of the reboot process
6. *however* in (4) sigterm is still blocked to systemd waits 1:30min

to fix this we need to stop listening to sigterm in (4). This PR
does that (thanks to Samuele) and adds one test for it. There is
one test missing, i.e. to test that snapd has stopped listening
to sigterm. This is indirect because the test harness seems
to also catch sigterm so we can't just setup a new signal handler
and send sigterm. So we just check if the sigCh got closed.