spdx: allow Other-Open-Source
We currently have a "Other Open Source" license in our supported
licenses. This license is not part of spdx and it was added for
store compatibility. However the spaces inside this name will
not work with the spdx parser and it looks also very out of place
with the other licenses.

This PR changes the license to Other-Open-Source so that the
parser can deal with it.  This will allow snaps to set
`license: Other-Open-Source` in their snap.yaml.

We will also make sure that the store is aware of this.