image: add type to seed.yaml to allow sorting in firstboot
This PR will allow us to sort by snap type in firstboot to
ensure that bases are installed before app snaps. No functional
changes yet except for adding the snap type to seed.yaml and
to update the tests.

Note that we could also do this differently and read the snap
info in the firstboot code and use that for sorting. If the
other approach is considered cleaner I'm happy to drop this PR
and do it using the other method.