snap: add new type "TypeSnapd" and attach to the snapd snap
We use types sort sorting but the snapd snap is handled special
(in the sort code and in more places). By adding a new "TypeSnapd"
snap type this becomes more uniform.

The current approach is to assign the type when reading the
snap.yaml. This is mostly to make the transition easier for
people who have snapd installed already but the snapd does
not have the right type yet. We could omit it given that we
have no real core18 systems in the wild yet and that snapd
is not installable yet on classic/core16 systems.