snapd: go into degraded mode when the selftest fails
If the selftest fails the daemon dies currently. This is not ideal
because the user will need to dig through journalctl to find the
reason for the failure.

This PR changes the behaviour of the daemon to go into "degraded"
mode instead. The daemon will run but will reply with an internal
error to all API commands.

This means that on a system with no squashfs support the result
will look like this:
```
$ snap install josm
error: selftest failed: cannot mount squashfs image using "squashfs": exit
       status 1
```
which means the user has an easier time to figure out what is going
on. Before we reported:
```
error: cannot communicate with server: Post http://localhost/v2/snaps/josm: dial unix /run/snapd.socket: connect: no such file or directory
```