interfaces: retain order of inserted security backends
This patch makes the interface repository retain the order of inserted
security backends. This in turns makes the interface manager respect
that order (as it uses the repository to know about available backends).
This in turn means that the order defined in the interfaces/backends/
package, specifically the All variable there, matters.

The order defined there is:
 - systemd
 - seccomp
 - dbus
 - udev
 - mount
 - kmod
 - apparmor (optionally)

The prior order is essentially the same list, sorted alphabetically:

 - apparmor (optionally)
 - dbus
 - kmod
 - mount
 - seccomp
 - systemd
 - udev

This means that the apparmor module would always go first. This was okay
for most interfaces but unfortunately the gpio interface depends on the
systemd backend (specifically it needs to read a file that only exists
after the systemd backend has completed its task).

Forum: https://forum.snapcraft.io/t/snap-enable-failure/6307/
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>