interfaces: make apparmor-features depend on AppArmor support level
It is possible that various helpers report non-empty AppArmor features list, but
the overall AppArmor support level is NoAppArmor due to missing userspace
tooling. In such case make sure to not write the features is AppArmor cannot be
used. When the userspace bits are avaialble at some point, system-key will
differ and relevant profiles will be regenerated.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>