snap: rename to ExpandSnapMountVariables(), expand in the context of snap's mount namespace
Rename ExpandSnapVariables() to ExpandSnapMountVariables() and have the function
resolve $SNAP, $SNAP_DATA, $SNAP_COMMON in the context of snap's mount
namespace. Meaning, the resolved values are the same for snap with instance key
and those without. In broader context, the low level mount namespace setup is
resposible for manipulating the snap's mount ns to ensure that snaps with
instance key have a system view nearly identical to snaps without instance key.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>