cmd,interfaces,tests: add /mnt to removable-media interface (#5307)
* cmd/snap-confine: allow hard-coded mounts to be optional

This patch changes the code path processing the list of built-in mount
operations so that certain operations are allowed to fail if the mount
point or mount source are missing.

This is convenient when making base snaps as we now have a vocabulary to
say that something is actually optional.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* cmd/snap-confine: optionally bind-mount /mnt

This patch puts the /mnt directory from the host into the mount
namespace of snap applications. Unlike other such entires this one is
marked as optional as our frozen-in-time test base snap doesn't carry
it.

This feature will be used as a foundation of the legacy-mnt interface.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* interfaces/builtin: add legacy-mnt interface

This patch adds a new interface for accessing mounted file systems under
the legacy /mnt directory. It is a relatively trivial interface, made
only slightly more complicated by a new entry in the list of static bind
mounts created by snap-confine.

Forum: https://forum.snapcraft.io/t//4669/
Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* interfaces/builtin: rename legacy-mnt to just mnt

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* tests: add snap and test for the mnt interface

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* tests: add mnt interface to the policy-app-consumer

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* cmd/snap-confine: add extra comment

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* many: merge mnt into removable-media interface

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* cmd/libsnap,cmd/snap-confine: add and use sc_do_optional_mount

This patch adds an variant of sc_do_mount that can silently fail with
ENOENT. The existing prior inline variant of the mount helper is now
replaced with a simple call to sc_do_optional_mount. The implementation
is now also shared between the optional and mandatory mount calls.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* tests: merge mnt tests into removable-media

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* cmd/snap-confine: if optional mount fails, don't change sharing mode

When an optional mount operation fails we now have a return value to
inspect that carries the success or failure information. When a optional
mount fails we no longer try to change sharing mode on the mount point.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* cmd/libsnap: add unit test for sc_do_optional_mount

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* tests: remove leftovers from separate interface

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* cmd/snap-confine: remove unused include

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* cmd/snap-confine: explain why /mnt is optional

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* interfaces: relax rules for removable-media

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>

* cmd/libsnap: extra tests for sc_do_optional_mount

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>