snapcraft: do not use --diry in mkversion (because LP:#1662388)
When running the snapd snap and using `snap version` the output
will currently show a version string with ".dirty" appended.

The reason for this is that snapcraft modifies the git tree
during building, see:
 https://bugs.launchpad.net/snapcraft/+bug/1662388

Until this bug is fixed we will need to generate the version
without the --dirty flag.