cmd: add systemd environment generator (#5736)
* cmd: add systemd environment generator

We currently only ship an environment generator for the user
sessions. However we need a environment generator that runs
at system startup time and creates environment available for
units that run outside of the user context (e.g. the
cloud-init-service). This PR adds such a system generator.

Systemd will not be able to pickup the system environment
generator changes via `systemctl daemon-reload` or `systemctl
daemon-reexec`. A full reboot is needed for this.