cmd/snap-update-ns: re-factor pair of helpers to call fstatfs once
This patch changes IsReadOnly and IsSnapdCreatedPrivateTmpfs to accept
a syscall.Statfs_t buffer and delegate the call that populates the
buffer to the upcoming helper that combines the two. I did this after
noticing a pattern of repeated fstatfs, on the same file descriptor,
inside tests that would exercise this code path. This lets use save
one of the calls.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>