cmd/snap-update-ns: detach Mk{Prefix,{File,Dir,Symlink{,All}}} from Secure
With the introduction of the restrictions type it make much better fit
than the initial (originally intendend for the same purpose) Secure
type. The secure type was meant to be the global state-keeping helper
but as the code evolved it became clear that we need a pair of state
keepers (global and local) and that it makes more sense to pass the
local state as an argument rather than use the global state as method
recepient.

The simple, untangled, story is that the Secure type keeps track of the
global state and can be asked to create a local state for a given path
while all the filesystem writing functions just take the local state
as an optional argument. The argument is optional because the local
state is the Restrictions type, which represents write restrictions,
if any apply, to some path.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>