cmd/snap-update-ns: lift write restrictions after making a new directory
This is somewhat tricky but IMO correct: when we successfully create
a new directory it must have been created in the root directory of
a tmpfs that is mounted somewhere by snapd OR restrictions were already
lifted.

Because such directory must for sure exist as a new tmpfs inode we can
lift write restrictions to the reminder of the path that is being
constructed.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>