cmd/snap-update-ns: allow writes to /tmp, /var/snap and /snap/$SNAP_NAME
We are almost ready to enable the restricted write mode but for that
we still need to white-list a few places that we are reasonably expected
to write as a part of normal operation. Those include the parent directory
of $SNAP_DATA and $SNAP_COMMON (/var/snap) as well as the parent directory
of $SNAP (/snap/$SNAP_NAME, since we are always on the inside of the mount
namespace).

The white-list allows us to say that we expect to make changes there
even if those would hit a filesystem that is not isolated inside the
per-snap mount namespace.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>