cmd/snap-update-ns: enable restriction checking for mount changes
Before creating an empty file, empty directory or making a symlink
we will now check which write restrictions apply to the path in question.

This will make the code report trespassing errors.  We will now treat
them the same way as we treated read only filesystem errors - by creating
a writable mimic in the offending directory so that further writes are
kept private the to mount namespace, inside a privatly mounted tmpfs.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>