store: refactor tests so that they work as store_test package
The tests in the store code are one of the few places left that
do not use the "$pkg_test" idiom. This caused me some confusion
so I did a (very mechnical) translation of the tests. This
probably should have a followup that refactors some bits and
moves bits into smaller and more targeted suites. But to make
reviews easier this tries to minimize the changes as much as
possible.