overlord/snapstate: fill request seed for store interaction
When sending requests to the store we need to ensure that 'instance-key' as used
in the actual request is stable across requests and restarts for every snap.
For snaps with instance key set, the store package builds the request's
'instance-key' as a hash with optionally provided salt/seed.

Make sure that we fill the seed when building the list of current snaps for
store requests.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>