selftest: add test to ensure selftest.checks is up-to-date
It was found in https://github.com/snapcore/snapd/pull/5768
that a selftest was added but not run. To prevent issues like
this in the future a test is added that intospects the functions
of the selftest package and warns if any check* function is added
without also adding it to the checks that need to be run.